test(eth): move TestEthBaseFee into eth_fee_history_test.go
