+++ itests/eth_fee_history_test.go
@@ -11,7 +11,9 @@ import (
 
 	"github.com/filecoin-project/go-jsonrpc"
 	"github.com/filecoin-project/go-state-types/abi"
+	"github.com/filecoin-project/go-state-types/big"
 
+	"github.com/filecoin-project/lotus/build/buildconstants"
 	"github.com/filecoin-project/lotus/chain/types"
 	"github.com/filecoin-project/lotus/chain/types/ethtypes"
 	"github.com/filecoin-project/lotus/itests/kit"
@@ -189,4 +191,25 @@ func TestEthFeeHistory(t *testing.T) {
 		json.Marshal([]interface{}{5, "10", &[]float64{}}),
 	).Assert(require.NoError))
 	require.NoError(err)
+}
+
+// TestEthBaseFee verifies that eth_baseFee returns the minimum base fee on an
+// idle test network. Empty blocks decay the base fee by 12.5% per block.
+// Starting from InitialBaseFee (100e6), the floor is reached after ~104 blocks.
+func TestEthBaseFee(t *testing.T) {
+	kit.QuietAllLogsExcept()
+
+	blockTime := 100 * time.Millisecond
+	client, _, ens := kit.EnsembleMinimal(t, kit.MockProofs(), kit.ThroughRPC())
+
+	ctx, cancel := context.WithTimeout(context.Background(), time.Minute)
+	defer cancel()
+
+	ens.InterconnectAll().BeginMining(blockTime)
+
+	client.WaitTillChain(ctx, kit.HeightAtLeast(110))
+
+	baseFee, err := client.EthBaseFee(ctx)
+	require.NoError(t, err)
+	require.Equal(t, ethtypes.EthBigInt(big.NewInt(buildconstants.MinimumBaseFee)), baseFee)
 }
